+++ examples/dart_examples/bin/timer.dart
@@ -4,12 +4,8 @@ main() {
   final timer = Duration(milliseconds: 2000).toSignal(cancelOnError: true);
 
   effect(() {
-    final on = timer();
-
-    if (on != null) {
-      print(on);
-      var dateTime = DateTime.fromMillisecondsSinceEpoch(on.millis);
-      print("${dateTime.hour}:${dateTime.minute}:${dateTime.second}");
-    }
+    final tick = timer();
+    final dateTime = DateTime.fromMillisecondsSinceEpoch(tick.millis);
+    print("$tick => ${dateTime.hour}:${dateTime.minute}:${dateTime.second}");
   });
 }