+++ packages/signals/lib/signals_flutter.dart
@@ -4,5 +4,5 @@ export 'signals.dart';
 export 'src/extensions/signal.dart';
 export 'src/extensions/value_listenable.dart';
 export 'src/extensions/value_notifier.dart';
-export 'src/widgets/watch.dart';
-export 'src/watch_signal.dart';
+export 'src/watch/widget.dart';
+export 'src/watch/extension.dart';